move state check to trigger_entity
