+++ tests/helpers/test_trigger_template_entity.py
@@ -221,121 +221,6 @@ async def test_trigger_template_availability_with_syntax_error(
     assert "Error rendering availability template for test.entity" in caplog.text
 
 
-async def test_template_state(hass: HomeAssistant) -> None:
-    """Test manual trigger template entity with a state."""
-    config = {
-        CONF_NAME: template.Template("test_entity", hass),
-        CONF_ICON: template.Template(_ICON_TEMPLATE, hass),
-        CONF_PICTURE: template.Template(_PICTURE_TEMPLATE, hass),
-        CONF_STATE: template.Template("{{ value == 'on' }}", hass),
-    }
-
-    class TestEntity(ManualTriggerEntity):
-        """Test entity class."""
-
-        extra_template_keys = (CONF_STATE,)
-
-        @property
-        def state(self) -> bool | None:
-            """Return extra attributes."""
-            return self._rendered.get(CONF_STATE)
-
-    entity = TestEntity(hass, config)
-    entity.entity_id = "test.entity"
-    variables = entity._template_variables_with_value(STATE_ON)
-    entity._process_manual_data(variables)
-    await hass.async_block_till_done()
-
-    assert entity.state == "True"
-    assert entity.icon == "mdi:on"
-    assert entity.entity_picture == "/local/picture_on"
-
-    variables = entity._template_variables_with_value(STATE_OFF)
-    entity._process_manual_data(variables)
-    await hass.async_block_till_done()
-
-    assert entity.state == "False"
-    assert entity.icon == "mdi:off"
-    assert entity.entity_picture == "/local/picture_off"
-
-
-async def test_bad_template_state(hass: HomeAssistant) -> None:
-    """Test manual trigger template entity with a state."""
-    config = {
-        CONF_NAME: template.Template("test_entity", hass),
-        CONF_ICON: template.Template(_ICON_TEMPLATE, hass),
-        CONF_PICTURE: template.Template(_PICTURE_TEMPLATE, hass),
-        CONF_STATE: template.Template("{{ x - 1 }}", hass),
-    }
-
-    class TestEntity(ManualTriggerEntity):
-        """Test entity class."""
-
-        extra_template_keys = (CONF_STATE,)
-
-        @property
-        def state(self) -> bool | None:
-            """Return extra attributes."""
-            return self._rendered.get(CONF_STATE)
-
-    entity = TestEntity(hass, config)
-    entity.entity_id = "test.entity"
-    variables = entity._template_variables({"x": 1})
-    entity._process_manual_data(variables)
-    await hass.async_block_till_done()
-
-    assert entity.available is True
-    assert entity.state == "0"
-    assert entity.icon == "mdi:off"
-    assert entity.entity_picture == "/local/picture_off"
-
-    variables = entity._template_variables_with_value(STATE_OFF)
-    entity._process_manual_data(variables)
-    await hass.async_block_till_done()
-
-    assert entity.available is False
-    assert entity.state is None
-    assert entity.icon is None
-    assert entity.entity_picture is None
-
-
-async def test_template_state_syntax_error(
-    hass: HomeAssistant, caplog: pytest.LogCaptureFixture
-) -> None:
-    """Test manual trigger template entity when state render fails."""
-    config = {
-        CONF_NAME: template.Template("test_entity", hass),
-        CONF_ICON: template.Template(_ICON_TEMPLATE, hass),
-        CONF_PICTURE: template.Template(_PICTURE_TEMPLATE, hass),
-        CONF_STATE: template.Template("{{ incorrect ", hass),
-    }
-
-    class TestEntity(ManualTriggerEntity):
-        """Test entity class."""
-
-        extra_template_keys = (CONF_STATE,)
-
-        @property
-        def state(self) -> bool | None:
-            """Return extra attributes."""
-            return self._rendered.get(CONF_STATE)
-
-    entity = TestEntity(hass, config)
-    entity.entity_id = "test.entity"
-    hass.states.async_set("test.entity", STATE_ON)
-    await entity.async_added_to_hass()
-
-    variables = entity._template_variables_with_value(STATE_ON)
-    entity._process_manual_data(variables)
-    await hass.async_block_till_done()
-
-    assert f"Error rendering {CONF_STATE} template for test.entity" in caplog.text
-
-    assert entity.state is None
-    assert entity.icon is None
-    assert entity.entity_picture is None
-
-
 async def test_attribute_order(
     hass: HomeAssistant, caplog: pytest.LogCaptureFixture
 ) -> None:
